Clean up some code and extract RetentionPolicyAssignment tests
